GD Muter Initial Release (Windows/Linux/Source)
No more manually moving files around needed - just point and click :)

Gun sounds aren't done yet, but this will allow you to mute/unmute the rest of the 'annoying' game sounds, as well as making the ClassNames patch an easy single click.

I used AI to help, since I haven't done much coding previously that wasn't in a mod framework - this was a bit of a learning curve!  I'll push out the full release in a day or two, with the gun sounds covered as well.